issue 1004 update role api
