+++ sqle/api/app.go
@@ -8,7 +8,6 @@ import (
 	"github.com/actiontech/sqle/sqle/api/cloudbeaver_wrapper"
 	"github.com/actiontech/sqle/sqle/api/controller"
 	v1 "github.com/actiontech/sqle/sqle/api/controller/v1"
-	v2 "github.com/actiontech/sqle/sqle/api/controller/v2"
 	sqleMiddleware "github.com/actiontech/sqle/sqle/api/middleware"
 	"github.com/actiontech/sqle/sqle/config"
 	_ "github.com/actiontech/sqle/sqle/docs"
@@ -96,12 +95,9 @@ func StartApi(net *gracenet.Net, exitChan chan struct{}, config config.SqleConfi
 
 		// role
 		v1Router.GET("/roles", DeprecatedBy(apiV2), AdminUserAllowed())
-		v2Router.GET("/roles", v2.GetRoles, AdminUserAllowed())
 		v1Router.GET("/role_tips", v1.GetRoleTips, AdminUserAllowed())
 		v1Router.POST("/roles", DeprecatedBy(apiV2), AdminUserAllowed())
-		v2Router.POST("/roles", v2.CreateRole, AdminUserAllowed())
 		v1Router.PATCH("/roles/:role_name/", DeprecatedBy(apiV2), AdminUserAllowed())
-		v2Router.PATCH("/roles/:role_name/", v2.UpdateRole, AdminUserAllowed())
 		v1Router.DELETE("/roles/:role_name/", v1.DeleteRole, AdminUserAllowed())
 
 		// instance
